@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "PyFunctional"
-version = "1.4.3"
+version = "1.5.0"
 description = "Package for creating data pipelines with chain functional programming"
 authors = ["Pedro Rodriguez <me@pedro.ai>"]
 maintainers = ["Pedro Rodriguez <me@pedro.ai>"]
@@ -8,43 +8,91 @@ readme = "README.md"
 homepage = "https://github.com/EntilZha/PyFunctional"
 repository = "https://github.com/EntilZha/PyFunctional"
 documentation = "https://docs.pyfunctional.pedro.ai/en/master/"
-keywords = ["functional", "pipeline", "chain"]
+keywords = ["functional", "pipeline", "collection", "chain", "linq", "parallel"]
 license = "MIT"
 classifiers = [
     'Development Status :: 5 - Production/Stable',
     'Intended Audience :: Developers',
     'License :: OSI Approved :: MIT License',
-    'Programming Language :: Python :: 3.6',
-    'Programming Language :: Python :: 3.7',
     'Programming Language :: Python :: 3.8',
+    'Programming Language :: Python :: 3.9',
+    'Programming Language :: Python :: 3.10',
+    'Programming Language :: Python :: 3.11',
+    'Programming Language :: Python :: 3.12',
+    'Programming Language :: Python :: 3.13',
     'Programming Language :: Python :: Implementation :: CPython',
     'Programming Language :: Python :: Implementation :: PyPy',
     'Natural Language :: English',
     'Operating System :: OS Independent',
-    'Topic :: Software Development :: Libraries :: Python Modules'
+    'Topic :: Software Development :: Libraries :: Python Modules',
 ]
-packages = [
-    { include = "functional" }
-]
+packages = [{ include = "functional" }]
 
 [tool.poetry.dependencies]
-python = "^3.6,>=3.6.1"
+python = "^3.9.0"
 dill = ">=0.2.5"
 tabulate = "<=1.0.0"
-pandas = {version = "^1.0.3", optional = true}
+pandas = { version = ">=1.0.3", optional = true }
 
 [tool.poetry.extras]
 all = ["pandas"]
 
-[tool.poetry.dev-dependencies]
-pypandoc = "^1.5"
-black = "^19.10b0"
-pytest = "^5.4.1"
-pylint = "^2.4.4"
-pytest-cov = "^2.8.1"
-codecov = "^2.0.22"
-coverage = "<5.0.0"
+[tool.poetry.group.dev.dependencies]
+black = "^23.1"
+pytest = "^7.3.1"
+pylint = "^3.3.3"
+pytest-cov = "^4.0.0"
+coverage = "^7.2.5"
+mypy = "^1.1.1"
+types-tabulate = "^0.9.0.3"
+pandas-stubs = "^2.0.3.230814"
 
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
+
+[tool.coverage.report]
+exclude_lines = [
+  "no cov",
+  "if __name__ == .__main__.:",
+  "if TYPE_CHECKING:",
+]
+
+# COMMON SETTINGS BELOW
+[tool.pyright]
+pythonVersion = "3.8"
+
+[tool.mypy]
+python_version = "3.8"
+check_untyped_defs = false
+disallow_any_generics = false
+disallow_incomplete_defs = false
+disallow_untyped_defs = false
+no_implicit_optional = true
+warn_redundant_casts = true
+warn_unused_ignores = false
+
+[tool.ruff.lint]
+select = [
+  "F",
+  "FA",
+  "TCH",
+  "UP",
+  "PL",
+]
+ignore = [
+  "UP008",
+  "UP028",
+  "PLR0913",
+  "PLR2004",
+  "PLR5501",
+]
+
+# https://github.com/astral-sh/ruff/issues/4368
+[tool.ruff.lint.per-file-ignores]
+"**/test/*.py" = [
+  "ALL",
+]
+"docs/**/*.py" = [
+  "ALL",
+]
