@@ -2,7 +2,7 @@ from functools import partial
 from functional.util import compose, parallelize
 
 
-class ExecutionStrategies(object):
+class ExecutionStrategies:
     """
     Enum like object listing the types of execution strategies.
     """
@@ -11,7 +11,7 @@ class ExecutionStrategies(object):
     PARALLEL = 1
 
 
-class ExecutionEngine(object):
+class ExecutionEngine:
     """
     Class to perform serial execution of a Sequence evaluation.
     """
@@ -23,7 +23,6 @@ class ExecutionEngine(object):
         :param transformations: Transformations to apply
         :return: Resulting sequence or value
         """
-        # pylint: disable=no-self-use
         result = sequence
         for transform in transformations:
             strategies = transform.execution_strategies
