@@ -1,16 +1,27 @@
+from __future__ import annotations
 import re
 import csv as csvapi
 import json as jsonapi
 import sqlite3 as sqlite3api
 import builtins
 
+from itertools import chain
+from typing import Any, SupportsIndex, TypeVar, overload
+from collections.abc import Iterable, Iterator
+import typing
+
 from functional.execution import ExecutionEngine, ParallelExecutionEngine
 from functional.pipeline import Sequence
 from functional.util import is_primitive, default_value
 from functional.io import get_read_function
 
+if typing.TYPE_CHECKING:
+    from _typeshed import SupportsRead
 
-class Stream(object):
+_T = TypeVar("_T")
+
+
+class Stream:
     """
     Represents and implements a stream which separates the responsibilities of Sequence and
     ExecutionEngine.
@@ -18,15 +29,35 @@ class Stream(object):
     An instance of Stream is normally accessed as `seq`
     """
 
-    def __init__(self, disable_compression=False, max_repr_items=100, no_wrap=None):
+    def __init__(
+        self,
+        disable_compression: bool = False,
+        max_repr_items: int = 100,
+        no_wrap: bool | None = None,
+    ) -> None:
         """
         Default stream constructor.
         :param disable_compression: Disable file compression detection
         :param no_wrap: default value of no_wrap for functions like first() or last()
         """
-        self.disable_compression = disable_compression
-        self.max_repr_items = max_repr_items
-        self.no_wrap = no_wrap
+        self.disable_compression: bool = disable_compression
+        self.max_repr_items: int = max_repr_items
+        self.no_wrap: bool | None = no_wrap
+
+    @overload
+    def __call__(
+        self,
+        args: list[_T] | tuple[_T] | Iterable[_T] | Sequence[_T],
+        no_wrap: bool | None = None,
+        **kwargs,
+    ) -> Sequence[_T]:
+        ...
+
+    @overload
+    def __call__(
+        self, *args: _T, no_wrap: bool | None = None, **kwargs
+    ) -> Sequence[_T]:
+        ...
 
     def __call__(self, *args, no_wrap=None, **kwargs):
         """
@@ -41,13 +72,26 @@ class Stream(object):
         :param args: Sequence to wrap
         :return: Wrapped sequence
         """
-        # pylint: disable=no-self-use
         engine = ExecutionEngine()
         return self._parse_args(
             args, engine, no_wrap=default_value(no_wrap, self.no_wrap, False)
         )
 
-    def _parse_args(self, args, engine, no_wrap=None):
+    def chain(
+        self, *args: Iterable[_T], no_wrap: bool | None = None, **kwargs
+    ) -> Sequence[_T]:
+        """
+        Create a Sequence chaining multiple iterators.
+        """
+        for arg in args:
+            if not isinstance(arg, Iterable):
+                raise TypeError("The type of arg should be iterator.")
+
+        return self(chain(*args), no_wrap=no_wrap, **kwargs)
+
+    def _parse_args(
+        self, args, engine: ExecutionEngine | None, no_wrap: bool | None = None
+    ):
         _no_wrap = default_value(no_wrap, self.no_wrap, False)
         if len(args) == 0:
             return Sequence(
@@ -90,15 +134,17 @@ class Stream(object):
                 no_wrap=_no_wrap,
             )
 
+    # pylint: disable=unknown-option-value
+    # pylint: disable=too-many-positional-arguments
     def open(
         self,
-        path,
-        delimiter=None,
-        mode="r",
-        buffering=-1,
-        encoding=None,
-        errors=None,
-        newline=None,
+        path: str,
+        delimiter: str | None = None,
+        mode: str = "r",
+        buffering: int = -1,
+        encoding: str | None = None,
+        errors: str | None = None,
+        newline: str | None = None,
     ):
         """
         Reads and parses input files as defined.
@@ -137,6 +183,16 @@ class Stream(object):
         else:
             return self("".join(list(file)).split(delimiter))
 
+    @overload
+    def range(self, stop: SupportsIndex, /) -> Sequence[int]:
+        ...
+
+    @overload
+    def range(
+        self, start: SupportsIndex, stop: SupportsIndex, step: SupportsIndex = ..., /
+    ) -> Sequence[int]:
+        ...
+
     def range(self, *args):
         """
         Alias to range function where seq.range(args) is equivalent to seq(range(args)).
@@ -149,7 +205,9 @@ class Stream(object):
         """
         return self(builtins.range(*args))  # pylint: disable=no-member
 
-    def csv(self, csv_file, dialect="excel", **fmt_params):
+    def csv(
+        self, csv_file: str | Iterator[str], dialect="excel", **fmt_params
+    ) -> Sequence[list[str]]:
         """
         Reads and parses the input of a csv stream or file.
 
@@ -177,15 +235,17 @@ class Stream(object):
         csv_input = csvapi.reader(input_file, dialect=dialect, **fmt_params)
         return self(csv_input).cache(delete_lineage=True)
 
+    # pylint: disable=unknown-option-value
+    # pylint: disable=too-many-positional-arguments
     def csv_dict_reader(
         self,
-        csv_file,
-        fieldnames=None,
-        restkey=None,
-        restval=None,
-        dialect="excel",
-        **kwds
-    ):
+        csv_file: str | Iterator[str],
+        fieldnames: typing.Sequence[str] | None = None,
+        restkey: str | None = None,
+        restval: str | Any | None = None,
+        dialect: str = "excel",
+        **kwds,
+    ) -> Sequence[dict[str | Any, str | Any]]:
         if isinstance(csv_file, str):
             file_open = get_read_function(csv_file, self.disable_compression)
             input_file = file_open(csv_file)
@@ -202,11 +262,11 @@ class Stream(object):
             restkey=restkey,
             restval=restval,
             dialect=dialect,
-            **kwds
+            **kwds,
         )
         return self(csv_input).cache(delete_lineage=True)
 
-    def jsonl(self, jsonl_file):
+    def jsonl(self, jsonl_file: str | Iterator[str]) -> Sequence[Any]:
         """
         Reads and parses the input of a jsonl file stream or file.
 
@@ -226,7 +286,9 @@ class Stream(object):
             input_file = jsonl_file
         return self(input_file).map(jsonapi.loads).cache(delete_lineage=True)
 
-    def json(self, json_file):
+    def json(
+        self, json_file: str | Iterator[str] | SupportsRead[str | bytes]
+    ) -> Sequence[tuple[str, Any]] | Sequence[Any]:
         """
         Reads and parses the input of a json file handler or file.
 
@@ -247,9 +309,9 @@ class Stream(object):
         if isinstance(json_file, str):
             file_open = get_read_function(json_file, self.disable_compression)
             input_file = file_open(json_file)
-            json_input = jsonapi.load(input_file)
+            json_input = jsonapi.load(input_file)  # pyright: ignore[reportArgumentType]
         elif hasattr(json_file, "read"):
-            json_input = jsonapi.load(json_file)
+            json_input = jsonapi.load(json_file)  # type: ignore
         else:
             raise ValueError(
                 "json_file must be a file path or implement the iterator interface"
