@@ -65,7 +65,7 @@ class TestPipeline(unittest.TestCase):
         self.assertEqual(repr(l), repr(self.seq(l)))
 
     def test_lineage_name(self):
-        f = lambda x: x
+        f = lambda x: x  # noqa: E731
         self.assertEqual(f.__name__, name(f))
         f = "test"
         self.assertEqual("test", name(f))
@@ -294,9 +294,8 @@ class TestPipeline(unittest.TestCase):
 
     def test_drop_while(self):
         l = [1, 2, 3, 4, 5, 6, 7, 8]
-        f = lambda x: x < 4
         expect = [4, 5, 6, 7, 8]
-        result = self.seq(l).drop_while(f)
+        result = self.seq(l).drop_while(lambda x: x < 4)
         self.assertIteratorEqual(expect, result)
         self.assert_type(result)
 
@@ -311,9 +310,8 @@ class TestPipeline(unittest.TestCase):
 
     def test_take_while(self):
         l = [1, 2, 3, 4, 5, 6, 7, 8]
-        f = lambda x: x < 4
         expect = [1, 2, 3]
-        result = self.seq(l).take_while(f)
+        result = self.seq(l).take_while(lambda x: x < 4)
         self.assertIteratorEqual(result, expect)
         self.assert_type(result)
 
@@ -342,18 +340,16 @@ class TestPipeline(unittest.TestCase):
         self.assertSetEqual(result.set(), set(expect))
 
     def test_map(self):
-        f = lambda x: x * 2
         l = [1, 2, 0, 5]
         expect = [2, 4, 0, 10]
-        result = self.seq(l).map(f)
+        result = self.seq(l).map(lambda x: x * 2)
         self.assertIteratorEqual(expect, result)
         self.assert_type(result)
 
     def test_select(self):
-        f = lambda x: x * 2
         l = [1, 2, 0, 5]
         expect = [2, 4, 0, 10]
-        result = self.seq(l).select(f)
+        result = self.seq(l).select(lambda x: x * 2)
         self.assertIteratorEqual(expect, result)
         self.assert_type(result)
 
@@ -368,17 +364,26 @@ class TestPipeline(unittest.TestCase):
         self.assertIteratorEqual(expect, result)
         self.assert_type(result)
 
+    def test_remove_none(self) -> None:
+        l = [None, -1, 0, None, 10]
+        expect = [-1, 0, 10]
+        s = self.seq(l)
+        result = s.remove_none()
+        self.assertIteratorEqual(expect, result)
+        self.assert_type(result)
+
     def test_filter(self):
-        f = lambda x: x > 0
         l = [0, -1, 5, 10]
         expect = [5, 10]
         s = self.seq(l)
-        result = s.filter(f)
+        result = s.filter(lambda x: x > 0)
         self.assertIteratorEqual(expect, result)
         self.assert_type(result)
 
     def test_where(self):
-        f = lambda x: x > 0
+        def f(x):
+            return x > 0
+
         l = [0, -1, 5, 10]
         expect = [5, 10]
         s = self.seq(l)
@@ -967,7 +972,7 @@ class TestPipeline(unittest.TestCase):
     def test_wrap_pandas(self):
         df1 = pandas.DataFrame({"name": ["name1", "name2"], "value": [1, 2]})
         df2 = pandas.DataFrame({"name": ["name1", "name2"], "value": [3, 4]})
-        result = seq([df1, df2]).reduce(lambda x, y: x.append(y))
+        result = seq([df1, df2]).reduce(lambda x, y: pandas.concat([x, y]))
         self.assertEqual(result.len(), 4)
         self.assertEqual(result[0].to_list(), ["name1", 1])
         self.assertEqual(result[1].to_list(), ["name2", 2])
@@ -1013,7 +1018,7 @@ class TestPipeline(unittest.TestCase):
         if self.seq is pseq:
             raise self.skipTest("pseq doesn't support functions with side-effects")
         calls = []
-        func = lambda x: calls.append(x)
+        func = calls.append
         result = self.seq(1, 2, 3).map(func).cache().map(lambda x: x).to_list()
         self.assertEqual(len(calls), 3)
         self.assertEqual(result, [None, None, None])
@@ -1074,22 +1079,22 @@ class TestExtend(unittest.TestCase):
 
         @extend
         def square(it):
-            return [i ** 2 for i in it]
+            return [i**2 for i in it]
 
         result = seq.range(100).square().list()
-        expected = [i ** 2 for i in range(100)]
+        expected = [i**2 for i in range(100)]
         self.assertEqual(result, expected)
 
         name = "PARALLEL_SQUARE"
 
         @extend(parallel=True, name=name)
         def square_parallel(it):
-            return [i ** 2 for i in it]
+            return [i**2 for i in it]
 
         result = seq.range(100).square_parallel()
         self.assertEqual(result.sum(), sum(expected))
         self.assertEqual(
-            repr(result._lineage), "Lineage: sequence -> extended[%s]" % name
+            repr(result._lineage), f"Lineage: sequence -> extended[{name}]"
         )
 
         @extend
@@ -1114,8 +1119,8 @@ class TestExtend(unittest.TestCase):
         expected = array.array("f", range(10))
         self.assertEqual(result, expected)
 
-        result = seq.range(10).map(lambda x: x ** 2).toarray()
-        expected = array.array("f", [i ** 2 for i in range(10)])
+        result = seq.range(10).map(lambda x: x**2).toarray()
+        expected = array.array("f", [i**2 for i in range(10)])
         self.assertEqual(result, expected)
 
         # a more complex example combining all above
