@@ -11,6 +11,7 @@ from itertools import (
 )
 import collections
 import types
+from collections.abc import Callable
 
 from functional.execution import ExecutionStrategies
 
@@ -24,7 +25,7 @@ Transformation = collections.namedtuple(
 CACHE_T = Transformation("cache", None, None)
 
 
-def name(function):
+def name(function: Callable):
     """
     Retrieve a pretty name for the function
     :param function: function to get name from
@@ -36,79 +37,79 @@ def name(function):
         return str(function)
 
 
-def map_t(func):
+def map_t(func: Callable):
     """
     Transformation for Sequence.map
     :param func: map function
     :return: transformation
     """
     return Transformation(
-        "map({0})".format(name(func)),
+        f"map({name(func)})",
         partial(map, func),
         {ExecutionStrategies.PARALLEL},
     )
 
 
-def select_t(func):
+def select_t(func: Callable):
     """
     Transformation for Sequence.select
     :param func: select function
     :return: transformation
     """
     return Transformation(
-        "select({0})".format(name(func)),
+        f"select({name(func)})",
         partial(map, func),
         {ExecutionStrategies.PARALLEL},
     )
 
 
-def starmap_t(func):
+def starmap_t(func: Callable):
     """
     Transformation for Sequence.starmap and Sequence.smap
     :param func: starmap function
     :return: transformation
     """
     return Transformation(
-        "starmap({})".format(name(func)),
+        f"starmap({name(func)})",
         partial(starmap, func),
         {ExecutionStrategies.PARALLEL},
     )
 
 
-def filter_t(func):
+def filter_t(func: Callable):
     """
     Transformation for Sequence.filter
     :param func: filter function
     :return: transformation
     """
     return Transformation(
-        "filter({0})".format(name(func)),
+        f"filter({name(func)})",
         partial(filter, func),
         {ExecutionStrategies.PARALLEL},
     )
 
 
-def where_t(func):
+def where_t(func: Callable):
     """
     Transformation for Sequence.where
     :param func: where function
     :return: transformation
     """
     return Transformation(
-        "where({0})".format(name(func)),
+        f"where({name(func)})",
         partial(filter, func),
         {ExecutionStrategies.PARALLEL},
     )
 
 
-def filter_not_t(func):
+def filter_not_t(func: Callable):
     """
     Transformation for Sequence.filter_not
     :param func: filter_not function
     :return: transformation
     """
     return Transformation(
-        "filter_not({0})".format(name(func)),
+        f"filter_not({name(func)})",
         partial(filterfalse, func),
         {ExecutionStrategies.PARALLEL},
     )
@@ -122,7 +123,7 @@ def reversed_t():
     return Transformation("reversed", reversed, [ExecutionStrategies.PRE_COMPUTE])
 
 
-def slice_t(start, until):
+def slice_t(start: int, until: int):
     """
     Transformation for Sequence.slice
     :param start: start index
@@ -130,7 +131,7 @@ def slice_t(start, until):
     :return: transformation
     """
     return Transformation(
-        "slice({0}, {1})".format(start, until),
+        f"slice({start}, {until})",
         lambda sequence: islice(sequence, start, until),
         None,
     )
@@ -153,7 +154,7 @@ def distinct_t():
     return Transformation("distinct", distinct, None)
 
 
-def distinct_by_t(func):
+def distinct_by_t(func: Callable):
     """
     Transformation for Sequence.distinct_by
     :param func: distinct_by function
@@ -168,10 +169,10 @@ def distinct_by_t(func):
                 distinct_lookup[key] = element
         return distinct_lookup.values()
 
-    return Transformation("distinct_by({0})".format(name(func)), distinct_by, None)
+    return Transformation(f"distinct_by({name(func)})", distinct_by, None)
 
 
-def sorted_t(key=None, reverse=False):
+def sorted_t(key=None, reverse: bool = False):
     """
     Transformation for Sequence.sorted
     :param key: key to sort by
@@ -183,20 +184,20 @@ def sorted_t(key=None, reverse=False):
     )
 
 
-def order_by_t(func):
+def order_by_t(func: Callable):
     """
     Transformation for Sequence.order_by
     :param func: order_by function
     :return: transformation
     """
     return Transformation(
-        "order_by({0})".format(name(func)),
+        f"order_by({name(func)})",
         lambda sequence: sorted(sequence, key=func),
         None,
     )
 
 
-def drop_right_t(n):
+def drop_right_t(n: int):
     """
     Transformation for Sequence.drop_right
     :param n: number to drop from right
@@ -207,57 +208,51 @@ def drop_right_t(n):
     else:
         end_index = -n
     return Transformation(
-        "drop_right({0})".format(n),
+        f"drop_right({n})",
         lambda sequence: sequence[:end_index],
         [ExecutionStrategies.PRE_COMPUTE],
     )
 
 
-def drop_t(n):
+def drop_t(n: int):
     """
     Transformation for Sequence.drop
     :param n: number to drop from left
     :return: transformation
     """
     return Transformation(
-        "drop({0})".format(n), lambda sequence: islice(sequence, n, None), None
+        f"drop({n})", lambda sequence: islice(sequence, n, None), None
     )
 
 
-def drop_while_t(func):
+def drop_while_t(func: Callable):
     """
     Transformation for Sequence.drop_while
     :param func: drops while func is true
     :return: transformation
     """
-    return Transformation(
-        "drop_while({0})".format(name(func)), partial(dropwhile, func), None
-    )
+    return Transformation(f"drop_while({name(func)})", partial(dropwhile, func), None)
 
 
-def take_t(n):
+def take_t(n: int):
     """
     Transformation for Sequence.take
     :param n: number to take
     :return: transformation
     """
-    return Transformation(
-        "take({0})".format(n), lambda sequence: islice(sequence, 0, n), None
-    )
+    return Transformation(f"take({n})", lambda sequence: islice(sequence, 0, n), None)
 
 
-def take_while_t(func):
+def take_while_t(func: Callable):
     """
     Transformation for Sequence.take_while
     :param func: takes while func is True
     :return: transformation
     """
-    return Transformation(
-        "take_while({0})".format(name(func)), partial(takewhile, func), None
-    )
+    return Transformation(f"take_while({name(func)})", partial(takewhile, func), None)
 
 
-def flat_map_impl(func, sequence):
+def flat_map_impl(func: Callable, sequence):
     """
     Implementation for flat_map_t
     :param func: function to map
@@ -265,8 +260,7 @@ def flat_map_impl(func, sequence):
     :return: flat_map generator
     """
     for element in sequence:
-        for value in func(element):
-            yield value
+        yield from func(element)
 
 
 def flat_map_t(func):
@@ -276,7 +270,7 @@ def flat_map_t(func):
     :return: transformation
     """
     return Transformation(
-        "flat_map({0})".format(name(func)),
+        f"flat_map({name(func)})",
         partial(flat_map_impl, func),
         {ExecutionStrategies.PARALLEL},
     )
@@ -472,7 +466,7 @@ def reduce_by_key_t(func):
     :return: transformation
     """
     return Transformation(
-        "reduce_by_key({0})".format(name(func)), partial(reduce_by_key_impl, func), None
+        f"reduce_by_key({name(func)})", partial(reduce_by_key_impl, func), None
     )
 
 
@@ -493,7 +487,7 @@ def accumulate_t(func):
     Transformation for Sequence.accumulate
     """
     return Transformation(
-        "accumulate({0})".format(name(func)), partial(accumulate_impl, func), None
+        f"accumulate({name(func)})", partial(accumulate_impl, func), None
     )
 
 
@@ -559,9 +553,7 @@ def group_by_t(func):
     :param func: grouping function
     :return: transformation
     """
-    return Transformation(
-        "group_by({0})".format(name(func)), partial(group_by_impl, func), None
-    )
+    return Transformation(f"group_by({name(func)})", partial(group_by_impl, func), None)
 
 
 def grouped_impl(size, sequence):
@@ -577,7 +569,7 @@ def grouped_impl(size, sequence):
             batch = islice(iterator, size)
             yield list(chain((next(batch),), batch))
     except StopIteration:
-        return
+        pass
 
 
 def grouped_t(size):
@@ -586,9 +578,7 @@ def grouped_t(size):
     :param size: size of groups
     :return: transformation
     """
-    return Transformation(
-        "grouped({0})".format(size), partial(grouped_impl, size), None
-    )
+    return Transformation(f"grouped({size})", partial(grouped_impl, size), None)
 
 
 def sliding_impl(wrap, size, step, sequence):
@@ -616,7 +606,7 @@ def sliding_t(wrap, size, step):
     :return: transformation
     """
     return Transformation(
-        "sliding({0}, {1})".format(size, step),
+        f"sliding({size}, {step})",
         partial(sliding_impl, wrap, size, step),
         {ExecutionStrategies.PRE_COMPUTE},
     )
@@ -642,7 +632,7 @@ def partition_t(wrap, func):
     :return: transformation
     """
     return Transformation(
-        "partition({0})".format(name(func)), partial(partition_impl, wrap, func), None
+        f"partition({name(func)})", partial(partition_impl, wrap, func), None
     )
 
 
@@ -704,7 +694,7 @@ def join_t(other, join_type):
     :return: transformation
     """
     return Transformation(
-        "{0}_join".format(join_type), partial(join_impl, other, join_type), None
+        f"{join_type}_join", partial(join_impl, other, join_type), None
     )
 
 
@@ -720,12 +710,10 @@ def peek_impl(func, sequence):
         yield element
 
 
-def peek_t(func):
+def peek_t(func: Callable):
     """
     Transformation for Sequence.peek
     :param func: peek function
     :return: transformation
     """
-    return Transformation(
-        "peek({0})".format(name(func)), partial(peek_impl, func), None
-    )
+    return Transformation(f"peek({name(func)})", partial(peek_impl, func), None)
