@@ -11,6 +11,7 @@
 #
 # All configuration values have a default; values that are commented out
 # serve to show the default.
+from __future__ import annotations
 
 import sys
 import os
@@ -48,9 +49,9 @@ source_suffix = ".rst"
 master_doc = "index"
 
 # General information about the project.
-project = u"PyFunctional"
-copyright = u"2019, Pedro Rodriguez"
-author = u"Pedro Rodriguez"
+project = "PyFunctional"
+copyright = "2019, Pedro Rodriguez"
+author = "Pedro Rodriguez"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -206,7 +207,7 @@ htmlhelp_basename = "PyFunctionaldoc"
 
 # -- Options for LaTeX output ---------------------------------------------
 
-latex_elements = {
+latex_elements: dict[str, str] = {
     # The paper size ('letterpaper' or 'a4paper').
     #'papersize': 'letterpaper',
     # The font size ('10pt', '11pt' or '12pt').
@@ -224,8 +225,8 @@ latex_documents = [
     (
         master_doc,
         "PyFunctional.tex",
-        u"PyFunctional Documentation",
-        u"Pedro Rodriguez",
+        "PyFunctional Documentation",
+        "Pedro Rodriguez",
         "manual",
     ),
 ]
@@ -255,7 +256,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [(master_doc, "pyfunctional", u"PyFunctional Documentation", [author], 1)]
+man_pages = [(master_doc, "pyfunctional", "PyFunctional Documentation", [author], 1)]
 
 # If true, show URL addresses after external links.
 # man_show_urls = False
@@ -270,7 +271,7 @@ texinfo_documents = [
     (
         master_doc,
         "PyFunctional",
-        u"PyFunctional Documentation",
+        "PyFunctional Documentation",
         author,
         "PyFunctional",
         "One line description of project.",
