@@ -2,12 +2,13 @@
 Package for creating data pipelines, LINQ-style data analysis, and functional programming. Imports
 the primary entrypoint at streams.seq
 """
-from functional.streams import seq, pseq
+
+from functional.streams import seq, pseq  # noqa: F401
 
 __author__ = "Pedro Rodriguez"
-__copyright__ = "Copyright 2021, Pedro Rodriguez"
+__copyright__ = "Copyright 2024, Pedro Rodriguez"
 __license__ = "MIT"
-__version__ = "1.4.3"
+__version__ = "1.5.0"
 __maintainer__ = "Pedro Rodriguez"
 __email__ = "me@pedro.ai"
 __status__ = "Production"
