@@ -3,12 +3,15 @@ import unittest
 import collections
 import sys
 import gzip
+from pathlib import Path
 from platform import system
 import lzma
 import bz2
 
 from functional import seq, pseq
 from functional.streams import Stream, ParallelStream
+
+project_root = Path(__file__).parent.parent.parent.absolute()
 
 
 class TestStreams(unittest.TestCase):
@@ -17,41 +20,46 @@ class TestStreams(unittest.TestCase):
         self.seq_c_disabled = Stream(disable_compression=True)
 
     def test_open(self):
-        with open("LICENSE.txt") as f:
+        file_name = f"{project_root}/LICENSE.txt"
+        with open(file_name, encoding="utf8") as f:
             data = f.readlines()
-        self.assertListEqual(data, self.seq.open("LICENSE.txt").to_list())
+        self.assertListEqual(data, self.seq.open(file_name).to_list())
 
         text = "".join(data).split(",")
-        self.assertListEqual(
-            text, self.seq.open("LICENSE.txt", delimiter=",").to_list()
-        )
+        self.assertListEqual(text, self.seq.open(file_name, delimiter=",").to_list())
 
         with self.assertRaises(ValueError):
-            self.seq.open("LICENSE.txt", mode="w").to_list()
+            self.seq.open(file_name, mode="w").to_list()
 
     def test_open_gzip(self):
         expect = ["line0\n", "line1\n", "line2"]
         self.assertListEqual(
             expect,
-            self.seq.open("functional/test/data/test.txt.gz", mode="rt").to_list(),
+            self.seq.open(
+                f"{project_root}/functional/test/data/test.txt.gz", mode="rt"
+            ).to_list(),
         )
 
     def test_open_bz2(self):
         expect = ["line0\n", "line1\n", "line2"]
         self.assertListEqual(
             expect,
-            self.seq.open("functional/test/data/test.txt.bz2", mode="rt").to_list(),
+            self.seq.open(
+                f"{project_root}/functional/test/data/test.txt.bz2", mode="rt"
+            ).to_list(),
         )
 
     def test_open_xz(self):
         expect = ["line0\n", "line1\n", "line2"]
         self.assertListEqual(
             expect,
-            self.seq.open("functional/test/data/test.txt.xz", mode="rt").to_list(),
+            self.seq.open(
+                f"{project_root}/functional/test/data/test.txt.xz", mode="rt"
+            ).to_list(),
         )
 
     def test_disable_compression(self):
-        file_name = "functional/test/data/test.txt.gz"
+        file_name = f"{project_root}/functional/test/data/test.txt.gz"
         with open(file_name, "rb") as f:
             expect = f.readlines()
         self.assertListEqual(
@@ -64,19 +72,73 @@ class TestStreams(unittest.TestCase):
         data = [-5, -3, -1, 1, 3, 5, 7]
         self.assertListEqual(data, self.seq.range(-5, 8, 2).to_list())
 
+    def test_lazyiness(self):
+        def yielder():
+            nonlocal step
+            step += 1
+            yield 1
+            step += 1
+            yield 2
+
+        step = 0
+        sequence = iter(seq(yielder()).map(str))
+        assert (
+            step == 0
+            and next(sequence) == "1"
+            and step == 1
+            and next(sequence) == "2"
+            and step == 2
+        )
+
+        step = 0
+        sequence = iter(seq.chain(yielder()).map(str))
+        assert (
+            step == 0
+            and next(sequence) == "1"
+            and step == 1
+            and next(sequence) == "2"
+            and step == 2
+        )
+
+    def test_chain(self):
+        data_a = range(1, 5)
+        data_b = range(6, 11)
+        self.assertEqual(
+            list(data_a) + list(data_b), self.seq.chain(data_a, data_b).to_list()
+        )
+
+        data_c = set(data_b)
+        self.assertEqual(
+            list(data_a) + list(data_c), self.seq.chain(data_a, data_c).to_list()
+        )
+
+        data_d = {"a": 1, "b": 2}
+        self.assertEqual(
+            list(data_a) + list(data_d.keys()), self.seq.chain(data_a, data_d).to_list()
+        )
+
+        self.assertEqual([], self.seq.chain().to_list())
+
+        with self.assertRaises(TypeError):
+            self.seq.chain(1, 2).to_list()
+
+        self.assertEqual(list(data_a), self.seq.chain(data_a).to_list())
+
+        self.assertEqual([1], self.seq.chain([1]).to_list())
+
     def test_csv(self):
-        result = self.seq.csv("functional/test/data/test.csv").to_list()
+        file_name = f"{project_root}/functional/test/data/test.csv"
+        result = self.seq.csv(file_name).to_list()
         expect = [["1", "2", "3", "4"], ["a", "b", "c", "d"]]
         self.assertEqual(expect, result)
-        with open("functional/test/data/test.csv", "r") as csv_file:
+        with open(file_name, "r", encoding="utf8") as csv_file:
             self.assertEqual(expect, self.seq.csv(csv_file).to_list())
         with self.assertRaises(ValueError):
             self.seq.csv(1)
 
     def test_csv_dict_reader(self):
-        result = self.seq.csv_dict_reader(
-            "functional/test/data/test_header.csv"
-        ).to_list()
+        file_name = f"{project_root}/functional/test/data/test_header.csv"
+        result = self.seq.csv_dict_reader(file_name).to_list()
         self.assertEqual(result[0]["a"], "1")
         self.assertEqual(result[0]["b"], "2")
         self.assertEqual(result[0]["c"], "3")
@@ -84,7 +146,7 @@ class TestStreams(unittest.TestCase):
         self.assertEqual(result[1]["b"], "5")
         self.assertEqual(result[1]["c"], "6")
 
-        with open("functional/test/data/test_header.csv", "r") as f:
+        with open(file_name, "r", encoding="utf8") as f:
             result = self.seq.csv_dict_reader(f).to_list()
         self.assertEqual(result[0]["a"], "1")
         self.assertEqual(result[0]["b"], "2")
@@ -97,28 +159,36 @@ class TestStreams(unittest.TestCase):
             self.seq.csv_dict_reader(1)
 
     def test_gzip_csv(self):
-        result = self.seq.csv("functional/test/data/test.csv.gz").to_list()
+        result = self.seq.csv(
+            f"{project_root}/functional/test/data/test.csv.gz"
+        ).to_list()
         expect = [["1", "2", "3", "4"], ["a", "b", "c", "d"]]
         self.assertEqual(expect, result)
         with self.assertRaises(ValueError):
             self.seq.csv(1)
 
     def test_bz2_csv(self):
-        result = self.seq.csv("functional/test/data/test.csv.bz2").to_list()
+        result = self.seq.csv(
+            f"{project_root}/functional/test/data/test.csv.bz2"
+        ).to_list()
         expect = [["1", "2", "3", "4"], ["a", "b", "c", "d"]]
         self.assertEqual(expect, result)
         with self.assertRaises(ValueError):
             self.seq.csv(1)
 
     def test_xz_csv(self):
-        result = self.seq.csv("functional/test/data/test.csv.xz").to_list()
+        result = self.seq.csv(
+            f"{project_root}/functional/test/data/test.csv.xz"
+        ).to_list()
         expect = [["1", "2", "3", "4"], ["a", "b", "c", "d"]]
         self.assertEqual(expect, result)
         with self.assertRaises(ValueError):
             self.seq.csv(1)
 
     def test_jsonl(self):
-        result_0 = self.seq.jsonl("functional/test/data/test.jsonl").to_list()
+        result_0 = self.seq.jsonl(
+            f"{project_root}/functional/test/data/test.jsonl"
+        ).to_list()
         expect_0 = [[1, 2, 3], {"a": 1, "b": 2, "c": 3}]
         self.assertEqual(expect_0, result_0)
         result_1 = self.seq.jsonl(["[1, 2, 3]", "[4, 5, 6]"])
@@ -126,35 +196,41 @@ class TestStreams(unittest.TestCase):
         self.assertEqual(expect_1, result_1)
 
     def test_gzip_jsonl(self):
-        result_0 = self.seq.jsonl("functional/test/data/test.jsonl.gz").to_list()
+        result_0 = self.seq.jsonl(
+            f"{project_root}/functional/test/data/test.jsonl.gz"
+        ).to_list()
         expect_0 = [[1, 2, 3], {"a": 1, "b": 2, "c": 3}]
         self.assertEqual(expect_0, result_0)
 
     def test_bz2_jsonl(self):
-        result_0 = self.seq.jsonl("functional/test/data/test.jsonl.bz2").to_list()
+        result_0 = self.seq.jsonl(
+            f"{project_root}/functional/test/data/test.jsonl.bz2"
+        ).to_list()
         expect_0 = [[1, 2, 3], {"a": 1, "b": 2, "c": 3}]
         self.assertEqual(expect_0, result_0)
 
     def test_xz_jsonl(self):
-        result_0 = self.seq.jsonl("functional/test/data/test.jsonl.xz").to_list()
+        result_0 = self.seq.jsonl(
+            f"{project_root}/functional/test/data/test.jsonl.xz"
+        ).to_list()
         expect_0 = [[1, 2, 3], {"a": 1, "b": 2, "c": 3}]
         self.assertEqual(expect_0, result_0)
 
     def test_json(self):
-        list_test_path = "functional/test/data/test_list.json"
-        dict_test_path = "functional/test/data/test_dict.json"
+        list_test_path = f"{project_root}/functional/test/data/test_list.json"
+        dict_test_path = f"{project_root}/functional/test/data/test_dict.json"
         list_expect = [1, 2, 3, 4, 5]
-        dict_expect = list({u"a": 1, u"b": 2, u"c": 3}.items())
+        dict_expect = list({"a": 1, "b": 2, "c": 3}.items())
 
         result = self.seq.json(list_test_path).to_list()
         self.assertEqual(list_expect, result)
         result = self.seq.json(dict_test_path).to_list()
         self.assertEqual(dict_expect, result)
 
-        with open(list_test_path) as file_handle:
+        with open(list_test_path, encoding="utf8") as file_handle:
             result = self.seq.json(file_handle).to_list()
             self.assertEqual(list_expect, result)
-        with open(dict_test_path) as file_handle:
+        with open(dict_test_path, encoding="utf8") as file_handle:
             result = self.seq.json(file_handle).to_list()
             self.assertEqual(dict_expect, result)
 
@@ -162,10 +238,10 @@ class TestStreams(unittest.TestCase):
             self.seq.json(1)
 
     def test_gzip_json(self):
-        list_test_path = "functional/test/data/test_list.json.gz"
-        dict_test_path = "functional/test/data/test_dict.json.gz"
+        list_test_path = f"{project_root}/functional/test/data/test_list.json.gz"
+        dict_test_path = f"{project_root}/functional/test/data/test_dict.json.gz"
         list_expect = [1, 2, 3, 4, 5]
-        dict_expect = list({u"a": 1, u"b": 2, u"c": 3}.items())
+        dict_expect = list({"a": 1, "b": 2, "c": 3}.items())
 
         result = self.seq.json(list_test_path).to_list()
         self.assertEqual(list_expect, result)
@@ -176,10 +252,10 @@ class TestStreams(unittest.TestCase):
             self.seq.json(1)
 
     def test_bz2_json(self):
-        list_test_path = "functional/test/data/test_list.json.bz2"
-        dict_test_path = "functional/test/data/test_dict.json.bz2"
+        list_test_path = f"{project_root}/functional/test/data/test_list.json.bz2"
+        dict_test_path = f"{project_root}/functional/test/data/test_dict.json.bz2"
         list_expect = [1, 2, 3, 4, 5]
-        dict_expect = list({u"a": 1, u"b": 2, u"c": 3}.items())
+        dict_expect = list({"a": 1, "b": 2, "c": 3}.items())
 
         result = self.seq.json(list_test_path).to_list()
         self.assertEqual(list_expect, result)
@@ -190,10 +266,10 @@ class TestStreams(unittest.TestCase):
             self.seq.json(1)
 
     def test_xz_json(self):
-        list_test_path = "functional/test/data/test_list.json.xz"
-        dict_test_path = "functional/test/data/test_dict.json.xz"
+        list_test_path = f"{project_root}/functional/test/data/test_list.json.xz"
+        dict_test_path = f"{project_root}/functional/test/data/test_dict.json.xz"
         list_expect = [1, 2, 3, 4, 5]
-        dict_expect = list({u"a": 1, u"b": 2, u"c": 3}.items())
+        dict_expect = list({"a": 1, "b": 2, "c": 3}.items())
 
         result = self.seq.json(list_test_path).to_list()
         self.assertEqual(list_expect, result)
@@ -204,7 +280,7 @@ class TestStreams(unittest.TestCase):
             self.seq.json(1)
 
     def test_sqlite3(self):
-        db_file = "functional/test/data/test_sqlite3.db"
+        db_file = f"{project_root}/functional/test/data/test_sqlite3.db"
 
         # test failure case
         with self.assertRaises(ValueError):
@@ -259,18 +335,18 @@ class TestStreams(unittest.TestCase):
             pass
 
     def test_to_file(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
         sequence = self.seq(1, 2, 3, 4)
         sequence.to_file(tmp_path)
-        with open(tmp_path, "r") as output:
+        with open(tmp_path, "r", encoding="utf8") as output:
             self.assertEqual("[1, 2, 3, 4]", output.readlines()[0])
 
         sequence.to_file(tmp_path, delimiter=":")
-        with open(tmp_path, "r") as output:
+        with open(tmp_path, "r", encoding="utf8") as output:
             self.assertEqual("1:2:3:4", output.readlines()[0])
 
     def test_to_file_compressed(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
         sequence = self.seq(1, 2, 3, 4)
         sequence.to_file(tmp_path, compression="gzip")
         with gzip.open(tmp_path, "rt") as output:
@@ -285,7 +361,7 @@ class TestStreams(unittest.TestCase):
             self.assertEqual("[1, 2, 3, 4]", output.readlines()[0])
 
     def test_to_jsonl(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
         elements = [{"a": 1, "b": 2}, {"c": 3}, {"d": 4}]
         sequence = self.seq(elements)
 
@@ -294,7 +370,7 @@ class TestStreams(unittest.TestCase):
         self.assertEqual(elements, result)
 
     def test_to_jsonl_compressed(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
         elements = [{"a": 1, "b": 2}, {"c": 3}, {"d": 4}]
         sequence = self.seq(elements)
 
@@ -311,23 +387,23 @@ class TestStreams(unittest.TestCase):
         self.assertEqual(elements, result)
 
     def test_to_json(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
-        elements = [[u"a", 1], [u"b", 2], [u"c", 3]]
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
+        elements = [["a", 1], ["b", 2], ["c", 3]]
         sequence = self.seq(elements)
 
         sequence.to_json(tmp_path)
         result = self.seq.json(tmp_path).to_list()
         self.assertEqual(elements, result)
 
-        dict_expect = {u"a": 1, u"b": 2, u"c": 3}
+        dict_expect = {"a": 1, "b": 2, "c": 3}
         sequence.to_json(tmp_path, root_array=False)
         result = self.seq.json(tmp_path).to_dict()
         self.assertEqual(dict_expect, result)
 
     def test_to_json_compressed(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
-        elements = [[u"a", 1], [u"b", 2], [u"c", 3]]
-        dict_expect = {u"a": 1, u"b": 2, u"c": 3}
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
+        elements = [["a", 1], ["b", 2], ["c", 3]]
+        dict_expect = {"a": 1, "b": 2, "c": 3}
         sequence = self.seq(elements)
 
         sequence.to_json(tmp_path, compression="gzip")
@@ -355,7 +431,7 @@ class TestStreams(unittest.TestCase):
         self.assertEqual(dict_expect, result)
 
     def test_to_csv(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
         elements = [[1, 2, 3], [4, 5, 6], ["a", "b", "c"]]
         expect = [["1", "2", "3"], ["4", "5", "6"], ["a", "b", "c"]]
         sequence = self.seq(elements)
@@ -365,7 +441,7 @@ class TestStreams(unittest.TestCase):
 
     @unittest.skipUnless(system().startswith("Win"), "Skip CSV test if not on Windows")
     def test_to_csv_win(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
         elements = [[1, 2, 3], [4, 5, 6], ["a", "b", "c"]]
         expect = [["1", "2", "3"], [], ["4", "5", "6"], [], ["a", "b", "c"], []]
         sequence = self.seq(elements)
@@ -374,7 +450,7 @@ class TestStreams(unittest.TestCase):
         self.assertNotEqual(expect, result)
 
     def test_to_csv_compressed(self):
-        tmp_path = "functional/test/data/tmp/output.txt"
+        tmp_path = f"{project_root}/functional/test/data/tmp/output.txt"
         elements = [[1, 2, 3], [4, 5, 6], ["a", "b", "c"]]
         expect = [["1", "2", "3"], ["4", "5", "6"], ["a", "b", "c"]]
         sequence = self.seq(elements)
@@ -398,7 +474,7 @@ class TestStreams(unittest.TestCase):
             self.seq(elements).to_sqlite3(1, insert_sql)
 
     def test_to_sqlite3_file(self):
-        tmp_path = "functional/test/data/tmp/test.db"
+        tmp_path = f"{project_root}/functional/test/data/tmp/test.db"
 
         with sqlite3.connect(tmp_path) as conn:
             conn.execute("DROP TABLE IF EXISTS user;")
